Initial Commit for libkcapi-sys Crate
Libkcapi is a very useful library written in C for using and
testing the Linux Kernel Crypto API (KCAPI). However, as of today
no upstream rust bindings exist for this library.

This repository shall hold the raw low-level rust bindings for
libkcapi. Therefore, this marks the creation of libkcapi-sys,
the raw rust-bindings for libkcapi.

Signed-off-by: Purushottam A. Kulkarni <puruk@protonmail.com>