build: Build libkcapi with autotools
Libkcapi is now checked out as a submodule and therefore needs
to build within libkcapi-sys tree. For this purpose, we will use
the autotools crate.

In addition to this, a few tweaks need to be made to wrapper.h
and the build script to allow libkcapi and our bindings to compile.

This change:

* Adds the autotools build-dependency
* Modifies wrapper.h to use the local "kcapi.h" header
* Updates the build script to use autotools for compiling libkcapi