Implement tests for the kcapi_aead_* APIs
The KCAPI AEAD functions will be used used by the kcapi high-level
bindings. Therefore, we must implement tests to ensure that we
can actually perform encryption and decryption using this API.

For this purpose, we will test the gcm(aes) algorithm for encrypting
and decrypting known values.

This change implements simple Known Answer Tests for:

* Encryption using the kcapi_aead_* APIs
* Decryption using the kcapi_aead_* APIs