Deny on clippy Warnings in Github CI
We want to deny (cause errors) in the CI pipeline when clippy
warnings are thrown against our low-level bindings. This is so
that we can catch issues with the generated bindings themselves
when we upgrade to newer versions of libkcapi.

This change turns on deny on warnings for clippy in github CI.