Rename Package to kcapi-sys and Version to 1.0.0
This commit marks the creation of the first official version of
the low-level bindings for libkcapi. This version does not have
any functional changes when compared to v0.1.0.

From henceforth, this crate will be referred to as:

	kcapi-sys v1.0.0

The biggest chagne(s) are:

* The package is renamed from libkcapi-sys to kcapi-sys
* This version of kcapi-sys builds from libkcapi v1.3.1
* libkcapi v1.3.1 is checked out as a submodule of this repo
* There is greater test coverage.

Besides these, functionally these bindings remain the same as
v0.1.0