Enable Compilation of Asym and KPP
The libkcapi submodule of kcapi-sys does not compile asymmetric
and Key-agreement Protocol Primitives (KPP) support by default.

Since we now have the ability to test these APIs, we can add
the compilation options necessary for enabling the compilation
of these components to kcapi-sys' build script.

This change adds that support.