build: Prevent Modifications outside of OUT_DIR
Currently, the libkcapi build modifies a lot of files outside
OUT_DIR due to the use of automake. Additionally we should also
set the Cargo.toml fields to specify:

	- license-file
	- description
	- homepage
	- repository
	- readme

Additionally in order to copy directories into OUT_DIR, we need
the fs_extra crate as a build-dependency as rust does not provide
a standard way to copy directories recursively.