libkcapi: Switch to Using smuellerDD/libkcapi
Now that our patches to libkcapi have been merged, it would make
sense to start using smuellerDD's fork of libkcapi so that we no
longer have to maintain a private fork. This should have been done
in the last release, but it was a mistake on my part.

Therefore, this change:

* Updates the libkcapi submodule to refer to smuellerDD's repo